Add library code that implements the maketx command.
